src/ui/ws/requirements.txt
websockets==15.0.1
python-dotenv==1.1.0
pika==1.3.2
